pętla w destroy

+++ app/controllers/submissions_controller.rb
@@ -50,7 +50,11 @@ class SubmissionsController < ApplicationController
 
   def destroy
     submission = Submission.find(params[:id])
-    submission.destroy
+    if submission.article_revision.review.exists?
+      redirect_to @submission, flash: {notice: "Należy usunąć powiązane recenzje"}
+    else 
+      submission.destroy
+    end
     redirect_to submissions_path
   end
